+++ utils-test-app/utils_test_app/tests/test_esi.py
@@ -324,7 +324,7 @@ class TestFetchEsiStatus(TestCase):
             status_code=504,
         )
         status = fetch_esi_status()
-        self.assertEqual(requests_mocker.call_count, 3)
+        self.assertEqual(requests_mocker.call_count, 4)
         self.assertFalse(status.is_online)
 
     @patch(MODULE_PATH + ".sleep", lambda x: None)
@@ -340,7 +340,7 @@ class TestFetchEsiStatus(TestCase):
             status_code=502,
         )
         status = fetch_esi_status()
-        self.assertEqual(requests_mocker.call_count, 3)
+        self.assertEqual(requests_mocker.call_count, 4)
         self.assertFalse(status.is_online)
 
     @patch(MODULE_PATH + ".sleep", lambda x: None)